CNN_encoder/requirements_txt.txt
tensorflow>=2.10.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=8.3.0
requests>=2.25.0
tqdm>=4.62.0
scikit-learn>=1.0.0
pathlib2>=2.3.0
